+++ src/cadqueryhelper/shape/regularPolygon.py
@@ -16,7 +16,14 @@ from cqmore.polygon import regularPolygon
 
 def regular_polygon(radius=5, sides=6, height=5):
     diameter = radius*2
-    work = cqm.Workplane().makePolygon(regularPolygon(nSides = sides, radius = radius)).extrude(height)
+    work = (
+        cqm.Workplane()
+        .makePolygon(regularPolygon(nSides = sides, radius = radius))
+    )
+
+    if height:
+        work = work.extrude(height)
+
     work = work.translate((0,0,-1*(height/2)))
 
     meta = {'type':'regularPolygon','height':height, 'length':diameter, 'width':diameter}